@@ -11,13 +11,13 @@ import { Comp_1_x_1_1_Component } from './comp-tree/comp-1-x-1-1.component';
 import { Comp_1_x_1_2_Component } from './comp-tree/comp-1-x-1-2.component';
 import { Comp_1_x_1_3_Component } from './comp-tree/comp-1-x-1-3.component';
 import { Comp_1_x_1_4_Component } from './comp-tree/comp-1-x-1-4.component';
-import { Comp_1_x_2_Component } from 'app/comp-tree/comp-1-x-2.component';
-import { Comp_1_x_2_1_Component } from 'app/comp-tree/comp-1-x-2-1.component';
+import { Comp_1_x_2_Component } from './comp-tree/comp-1-x-2.component';
+import { Comp_1_x_2_1_Component } from './comp-tree/comp-1-x-2-1.component';
 import { Comp_1_x_2_2_Component } from './comp-tree/comp-1-x-2-2.component';
 import { Comp_1_x_2_3_Component } from './comp-tree/comp-1-x-2-3.component';
 import { Comp_1_x_2_4_Component } from './comp-tree/comp-1-x-2-4.component';
 import { Comp_1_x_3_Component } from './comp-tree/comp-1-x-3.component';
-import { Comp_1_x_3_1_Component } from 'app/comp-tree/comp-1-x-3-1.component';
+import { Comp_1_x_3_1_Component } from './comp-tree/comp-1-x-3-1.component';
 import { Comp_1_x_3_2_Component } from './comp-tree/comp-1-x-3-2.component';
 import { Comp_1_x_3_3_Component } from './comp-tree/comp-1-x-3-3.component';
 import { Comp_1_x_3_4_Component } from './comp-tree/comp-1-x-3-4.component';
